reorgnize keypoints_sstats and object_detection_stats
